chore(kms-connector): test s3 ct retrieval
