+++ kms-connector/tests/src/setup/mod.rs
@@ -1,0 +1,8 @@
+mod db;
+mod gw;
+mod instance;
+mod s3;
+
+pub use gw::*;
+pub use instance::*;
+pub use s3::*;
